pyproject.toml
[tool.poetry]
name = "ui-testing"
version = "0.1.0"
description = ""
authors = ["chevchelios <keks1001@mail.ru>"]
readme = "README.md"
packages = [{include = "ui_testing"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.1"
selene = "^2.0.0.b16"
python-dotenv = "^1.0.0"
allure-pytest = "^2.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = 'INFO'
log_cli_format = '%(asctime)s %(message)s'
log_cli_date_format = '%Y-%m-%d %H:%M:%S'
addopts = ' --alluredir=allure-results '
